Added JS templating and fixed the last modified tweaks
For people wanting to create Mustache, handlebar.js, handlebars.js and other flavours of javascript templates and have them compressedas a single asset, this is for you. Simply add an array to minify and it reads all the files, combines them into a Javascript object, assigns them to a variable, adds any postprocessing javascript and serves it up.

For the last modified headers, I've made some improvements/bugfixes. It should now be able to speedily exit when it detects a valid HTTP_IF_MODIFIED_SINCE header. I'm not bothering with etags as one meter is enough to determine if it has been changed.