Changed the way that build arrays are structured. Updated readme to reflect changes
The structure to indicate a template array was iffy. I've updated it so
that a template array is actually nested within an array. Here's how it
used to be:

    array(
        'item.js',
        'anotheritem.js',
        'templateVar' => array(
            'items.js',
            ...
        )
    )

This method of allowing mixed array keys, while nice in PHP, isn't
common in other languages or data formats (eg. my
[JsonReader](https://github.com/connrs/JsonReader) plugin) and so here
is the new format:

    array(
        'item.js',
        'anotheritem.js',
        array(
            'templateVar' => array(
                'item' => 'items.js',
                ...
            )
        )
    )

It adds extra weight in some ways but it also means a Json config file
can contain:

    [
        "item.js",
        "anotheritem.js",
        {"templateVar": [
                "item": "items.js",
                "item2": "b.js"
        ]}
    ]